+++ scripts/dev/02-gunicorn-server.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-gunicorn --worker-class eventlet --bind 0.0.0.0:7890  mmpm.wsgi:app --reload
+gunicorn --worker-class gevent --bind 0.0.0.0:7890  mmpm.wsgi:app --reload